Add precompiled viewed.res support
CI won't have TF2 installed

+++ dev/no_tutorial_viewed.sh
@@ -3,9 +3,13 @@
 BINDIR=$(dirname "$(readlink -fn "$0")")
 cd "${BINDIR}" || exit 2
 . ./mastercomfig-vars
-
-cp -f ./generate_viewed.sh "${TF2_DIR}/tf/maps"
-chmod +x "${TF2_DIR}/tf/maps/generate_viewed.sh"
-"${TF2_DIR}/tf/maps/generate_viewed.sh"
-mv -f "${TF2_DIR}/tf/maps/viewed.res" "../config/addons/mastercomfig-no-tutorial-addon/media"
-rm -f "${TF2_DIR}/tf/maps/generate_viewed.sh"
+if [ -z ${CI+x} ]
+  then
+    cp -f ./generate_viewed.sh "${TF2_DIR}/tf/maps"
+    chmod +x "${TF2_DIR}/tf/maps/generate_viewed.sh"
+    "${TF2_DIR}/tf/maps/generate_viewed.sh"
+    mv -f "${TF2_DIR}/tf/maps/viewed.res" "../config/addons/mastercomfig-no-tutorial-addon/media"
+    rm -f "${TF2_DIR}/tf/maps/generate_viewed.sh"
+  else
+    cp viewed.res "../config/addons/mastercomfig-no-tutorial-addon/media"
+fi
